Fixes as per review by @yorinasub17: - Moved mock tests in to separate files - fixed confusing loop in GetDefaultVpcs - Renamed several functions for clarity
